Add anchors to contributors regex
`NotProgramFOX123 =~ /ProgramFOX/i` does not return `nil` (which is what we need here), but it does for `/^ProgramFOX$/i`, so I added the anchors to be sure that GHULS doesn't match too many users. For most users, this would not have caused a problem, but for users with very short names it could be that their name is part of a name of another organization member.